src/dom_tree/handler.rs: started TreeNodeHandler
